// src/shared/api/models/getApiCashDrawerEmployeeSummary200Data.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { GetApiCashDrawerEmployeeSummary200DataCashDrawerId } from './getApiCashDrawerEmployeeSummary200DataCashDrawerId';
import type { GetApiCashDrawerEmployeeSummary200DataLastCashDrawerEndDate } from './getApiCashDrawerEmployeeSummary200DataLastCashDrawerEndDate';
import type { GetApiCashDrawerEmployeeSummary200DataDrawerName } from './getApiCashDrawerEmployeeSummary200DataDrawerName';
import type { GetApiCashDrawerEmployeeSummary200DataDrawerStatus } from './getApiCashDrawerEmployeeSummary200DataDrawerStatus';
import type { GetApiCashDrawerEmployeeSummary200DataDrawerStartDate } from './getApiCashDrawerEmployeeSummary200DataDrawerStartDate';
import type { GetApiCashDrawerEmployeeSummary200DataDrawerEndDate } from './getApiCashDrawerEmployeeSummary200DataDrawerEndDate';
import type { GetApiCashDrawerEmployeeSummary200DataPaymentBreakdown } from './getApiCashDrawerEmployeeSummary200DataPaymentBreakdown';
import type { GetApiCashDrawerEmployeeSummary200DataSalesBreakdown } from './getApiCashDrawerEmployeeSummary200DataSalesBreakdown';
import type { GetApiCashDrawerEmployeeSummary200DataCreditBreakdown } from './getApiCashDrawerEmployeeSummary200DataCreditBreakdown';
import type { GetApiCashDrawerEmployeeSummary200DataDetailedPaymentBreakdownItem } from './getApiCashDrawerEmployeeSummary200DataDetailedPaymentBreakdownItem';

export type GetApiCashDrawerEmployeeSummary200Data = {
    outletId: number;
    cashDrawerId: GetApiCashDrawerEmployeeSummary200DataCashDrawerId;
    fromDate: string;
    toDate: string;
    totalSales: number;
    totalTenderAmount: number;
    openingCash: number;
    closingCash: number;
    lastCashDrawerEndDate: GetApiCashDrawerEmployeeSummary200DataLastCashDrawerEndDate;
    drawerName: GetApiCashDrawerEmployeeSummary200DataDrawerName;
    drawerStatus: GetApiCashDrawerEmployeeSummary200DataDrawerStatus;
    drawerStartDate: GetApiCashDrawerEmployeeSummary200DataDrawerStartDate;
    drawerEndDate: GetApiCashDrawerEmployeeSummary200DataDrawerEndDate;
    paymentBreakdown: GetApiCashDrawerEmployeeSummary200DataPaymentBreakdown;
    salesBreakdown: GetApiCashDrawerEmployeeSummary200DataSalesBreakdown;
    creditBreakdown: GetApiCashDrawerEmployeeSummary200DataCreditBreakdown;
    detailedPaymentBreakdown?: GetApiCashDrawerEmployeeSummary200DataDetailedPaymentBreakdownItem[];
};